refactor: Utilizando o mesmo template para main-news e news-list
