+++ web/assets/js/utils/news-functions.js
@@ -1,4 +1,5 @@
 import { loadTemplate } from "./dom-functions.js";
+import { formatDateByTime, formatTimeByMs } from "./dates-functions.js";
 
 export function prepareLink(newsContent, newsItem) {
     // trata os links
@@ -65,4 +66,54 @@ function prepareCloseModalEvent(modalContent, modalElement) {
 export function removeSkeleton(mainElement, querySelectorName) {
     const skeleton = mainElement.querySelector(querySelectorName);
     skeleton.remove()
+}
+
+
+export async function prepareNewsElement(newsContent, newsItem) {
+    try {
+        const templateText = await loadTemplate('./assets/templates/news-item.html');
+        newsContent.innerHTML = templateText;
+
+        // trata a imagem
+        const image = newsContent.querySelector('.new-list-item__img-tag');
+        image.src = newsItem.image ? newsItem.image : './assets/imgs/G1-placeholder.webp';
+        image.alt = newsItem.title;
+
+        // trata o figcaption
+        const figcaptionImage = newsContent.querySelector('figcaption');
+        figcaptionImage.textContent = newsItem.title;
+
+        // trata as informações da noticia
+        const label = newsContent.querySelector('.new-list-item__label');
+        label.textContent = newsItem.section;
+
+        const title = newsContent.querySelector('.new-list-item__title');
+        title.textContent = newsItem.title;
+
+        const description = newsContent.querySelector('.new-list-item__description');
+        description.textContent = newsItem.summary;
+
+        const createAt = newsContent.querySelectorAll('.new-list-item__when');
+        createAt.forEach(when => {
+            when.textContent = formatDateByTime(newsItem.created);
+        })
+
+        const player = newsContent.querySelector('.new-list-item__play');
+        player.style.display = 'none';
+
+        prepareLink(newsContent, newsItem);
+
+
+        if (newsItem.video) {
+            image.classList.add('new-list-item__img-tag--video')
+            player.style.display = 'flex';
+            const playerTime = newsContent.querySelector('.new-list-item__play-time');
+            playerTime.textContent = formatTimeByMs(newsItem.video.duration);
+        }
+    } catch (error) {
+        console.log(error)
+        newsContent.textContent = "Erro ao carregar as noticias, por favor tente novamente mais tarde.";
+    }
+
+    return newsContent;
 }
